fix: 메인 페이지 section 에러를 위한 기술블로그 데이터 fetching 로직 컴포넌트 DynamicTechBlogComponent 내부로 이동
